Add anonymous shared name GUID to avoid kernel cache
Fixes #19671
We generate unique shared names for variables in eager mode to avoid
unintended sharing. This results in a memory leak due to the kernel
cache expanding with each new shared_name encountered. To avoid this
we reserve a shared name to send in with the op. When the handle
creation code sees this shared name, it generates a unique name
internally and returns a fresh handle each time.

PiperOrigin-RevId: 227932460