Update protobuf to 3.6.1 in install_proto3.sh
This fix updates protobuf to 3.6.1 in install_proto3.sh.

The protobuf library in workspace.bzl has been updated to 3.6.1 for
some time, though in CI build protoc still uses 3.6.0.

When build a custom op (in docker image `tensorflow/tensorflow:custom-op`)
and using grpc's `grpc_defs()` (`grpc/bazel/grpc_deps.bzl`):
https://github.com/grpc/grpc/blob/ac6795a57e05523b8fa220bc5cef26abb876aae5/bazel/grpc_deps.bzl#L5-L7
The issue arise because grpc's dependency is 3.6.1 (and an immediate earlier version is 3.5)
so there is no match to fit grpc+protobuf no matter which version of grpc is used.

This may not a problem for tensorflow itself as tensorflow does not call `grpc_defs()`. Tensorflow
links to protobuf 3.6.1 explicitly. Still, it would be good to match the
version of installed protoc (3.6.0) with protobuf library (3.6.1).

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>